version 1.1.10
- edit functionality almost done (still I have to add some logic and improve the design, but it's 90% done)

- fixed bug in url path, distance_between_prescriptions was undefined and it was a name conflict, I have 2 different names and I mistakenly used them wrong (I should use the same name)

+++ src/views/MoreDetailsRecipe.vue
@@ -5,6 +5,7 @@ export default {
   mounted() {
     let recipes = localStorage.getItem('recipes')
     this.recipes_list = recipes === null ? [] : JSON.parse(recipes)
+    console.log(this.$route.params.distance_between_recipes)
   },
   data() {
     return {
@@ -21,9 +22,10 @@ export default {
         recipe_duration: this.$route.params.recipe_duration,
         last_prescription_dates: this.$route.params.last_prescription_dates.split(','),
         status: this.$route.params.status,
-        distance_between_prescriptions: this.$route.params.distance_between_prescriptions,
+        distance_between_prescriptions: this.$route.params.distance_between_recipes,
       },
       recipes_list: [],
+      showModal: false,
     }
   },
   methods: {
@@ -46,22 +48,60 @@ export default {
           this.$route.params.doctor_name === this.recipes_list[r].doctor.name &&
           this.$route.params.recipe_duration === this.recipes_list[r].recipe_duration
         ) {
-          this.recipes_list[r].recipe_duration.split(" ");
+          this.recipes_list[r].recipe_duration.split(' ')
           let unformatted_current_date = new Date(Date.now())
-          this.recipes_list[r].current_prescription_date = `${unformatted_current_date.getFullYear()}-${unformatted_current_date.getMonth() + 1}-${unformatted_current_date.getDate()}`
-          let unformatted_future_date = new Date(Date.now() + (this.recipes_list[r].recipe_duration[0] * 1000 * 60 * 60 * 24 * 30))
-          this.recipes_list[r].future_prescription_date = `${unformatted_future_date.getFullYear()}-${unformatted_future_date.getMonth() + 1}-${unformatted_future_date.getDate()}`
+          this.recipes_list[r].current_prescription_date =
+            `${unformatted_current_date.getFullYear()}-${unformatted_current_date.getMonth() + 1}-${unformatted_current_date.getDate()}`
+          let unformatted_future_date = new Date(
+            Date.now() + this.recipes_list[r].recipe_duration[0] * 1000 * 60 * 60 * 24 * 30,
+          )
+          this.recipes_list[r].future_prescription_date =
+            `${unformatted_future_date.getFullYear()}-${unformatted_future_date.getMonth() + 1}-${unformatted_future_date.getDate()}`
 
-          this.recipes_list[r].last_prescription_dates.push(this.recipes_list[r].current_prescription_date)
+          this.recipes_list[r].last_prescription_dates.push(
+            this.recipes_list[r].current_prescription_date,
+          )
           this.recipe = this.recipes_list[r]
-          localStorage.setItem("recipes", JSON.stringify(this.recipes_list))
+          localStorage.setItem('recipes', JSON.stringify(this.recipes_list))
           this.goToRecipesView()
-          break;
+          break
         }
       }
     },
-    goToRecipesView(){
+    goToRecipesView() {
       this.$router.push(`/recipes`)
+    },
+    showConfirmDeletePopUp() {
+      this.showModal = !this.showModal;
+    },
+    deleteRecipe(){
+      let index = this.searchRecipe()
+      if (index > -1) {
+        this.recipes_list.splice(index, 1)
+        localStorage.setItem('recipes', JSON.stringify(this.recipes_list))
+        this.goToRecipesView()
+      }
+    },
+    searchRecipe() {
+      for (let i = 0; i < this.recipes_list.length; i++) {
+        if (this.recipes_list[i].patient.name === this.recipe.patient.name &&
+          this.recipes_list[i].doctor.name === this.recipe.doctor.name) {
+          return i;
+        }
+      }
+    },
+    goToEditRecipe() {
+      this.$router.push(
+        `/edit_recipe/${this.recipe.patient.name}/${this.recipe.doctor.name}/${this.recipe.doctor.specialization}/${this.recipe.recipe_duration}/${this.recipe.distance_between_prescriptions}/${this.recipe.last_prescription_dates}/${this.recipe.current_prescription_date}/${this.recipe.future_prescription_date}/${this.recipe.status}`,
+      )
+    },
+    updateRecipeDetails() {
+      for (let r in this.recipes_list) {
+        if (this.recipes_list[r].patient.name === this.recipe.patient.name &&
+          this.recipes_list[r].doctor.name === this.recipe.doctor.name) {
+
+        }
+      }
     },
   },
 }
@@ -70,35 +110,64 @@ export default {
   <div class="more-details-recipe">
     <div class="upper-zone">
       <div class="upper-zone-left">
-        <p>Pacient: {{ $route.params.patient_name }}</p>
-        <p>Tip rețetă: {{ $route.params.doctor_specialization }}</p>
-        <p>Doctor: {{ $route.params.doctor_name }}</p>
-        <p>Durata: {{ $route.params.recipe_duration }}</p>
+        <p>Pacient: {{ this.recipe.patient.name }}</p>
+        <p>Tip rețetă: {{ this.recipe.doctor.specialization }}</p>
+        <p>Doctor: {{ this.recipe.doctor.name }}</p>
+        <p>Durata: {{ this.recipe.recipe_duration }}</p>
       </div>
       <div class="upper-zone-right">
-        <button class="button-update" @click="calculateFuturePrescriptionDate" v-show="$route.params.status === '1'">Update</button>
-        <button class="button-delete">Delete</button>
-        <button class="button-edit">Edit</button>
+        <button
+          class="button-update"
+          @click="calculateFuturePrescriptionDate"
+          v-show="this.recipe.status === '1'"
+        >
+          Update
+        </button>
+        <button class="button-delete" @click="showConfirmDeletePopUp">Delete</button>
+        <button class="button-edit" @click="goToEditRecipe">Edit</button>
       </div>
     </div>
     <div class="middle-zone">
-      <p>Data ultimei rețete prescrise: {{ $route.params.current_prescription_date }}</p>
+      <p>Data ultimei rețete prescrise: {{ this.$route.params.current_prescription_date }}</p>
       <p>
-        Data următoarei prescrieri: {{ this.formatDate($route.params.future_prescription_date) }}
+        Data următoarei prescrieri: {{ this.formatDate(this.$route.params.future_prescription_date) }}
       </p>
       <p>Rețetele prescrise anterior:</p>
     </div>
     <div class="table-last-dates">
-      <p v-for="(index,counter) in recipe.last_prescription_dates" :key="index"> {{counter+1}}. {{index}}</p>
+      <p v-for="(index, counter) in this.recipe.last_prescription_dates" :key="index">
+        {{ counter + 1 }}. {{ index }}
+      </p>
+    </div>
+    <button @click="goToRecipesView">Back</button>
+  </div>
+
+  <div id="delete-confirmation-popUp" class="modal" v-show="showModal === true">
+    <div class="modal-content">
+      <h2 style="background-color: firebrick; color: white; font-weight: bold;">
+        Confirmare ștergere rețetă
+      </h2>
+      <p style="margin: 20px 2px 20px 2px; color: black">
+        Ești sigur că dorești să ștergi această rețetă?
+      </p>
+
+      <button class="button-cancel" @click="showConfirmDeletePopUp">Anulează</button>
+      <button class="button-confirm-deletion" @click="deleteRecipe">Șterge</button>
     </div>
   </div>
+
 </template>
 
 <style scoped>
+* {
+  box-sizing: border-box;
+}
 .more-details-recipe {
   background-color: #fff;
   color: #181818;
   padding: 10px 15px;
+  width: 100%;
+  height: 100%;
 }
 .table-last-dates {
   height: 100px;
@@ -119,12 +188,9 @@ export default {
   text-align: right;
 }
 
-.button-update,
-.button-delete,
-.button-edit {
-  display: block;
-  width: 70px;
-  height: 30px;
+button {
+  width: 80px;
+  height: 33px;
   margin-top: 3px;
   margin-right: 20px;
   margin-bottom: 6px;
@@ -135,13 +201,52 @@ export default {
   font-size: 13px;
   border-radius: 3px;
   border: none;
+  opacity: 0.9;
+}
+
+button:hover {
+  opacity: 1;
 }
 
 .button-edit {
+  display: block;
   background-color: forestgreen;
 }
 
 .button-delete {
+  display: block;
   background-color: firebrick;
+}
+
+.button-cancel {
+  background-color: darkgray;
+  margin-right: 10px;
+  margin-left: 10px;
+  margin-bottom: 20px;
+}
+
+.button-confirm-deletion {
+  background-color: firebrick;
+  margin-bottom: 20px;
+}
+
+#delete-confirmation-popUp {
+  text-align: center;
+  /*display: none; Hidden by default*/
+  position: fixed; /* Stay in place */
+  z-index: 1; /* Sit on top */
+  left: 0;
+  top: 0;
+  width: 100%; /* Full width */
+  height: 100%; /* Full height */
+  overflow: auto; /* Enable scroll if needed */
+  padding-top: 55%;
+}
+
+.modal-content {
+  background-color: #fefefe;
+  margin: 5% auto 15% auto; /* 5% from the top, 15% from the bottom and centered */
+  border: 1px solid black;
+  width: 80%; /* Could be more or less, depending on screen size */
 }
 </style>
